# docs/api/openapi.yaml
openapi: 3.2.0
...
info:
  title: Imperio Industrial — API pública
  version: 1.5.0
  summary: API REST del MMO de simulación económica, industrial y logística "Imperio Industrial".
  description: |
    API pública del servidor autoritativo de **Imperio Industrial** (GDD/SAD v1.2).

    ## Principios del contrato de API

    - **Una sola API para humanos y bots** (igualdad de API literal, ADR-010): mismos
      endpoints y mismos rate limits lógicos. El motor no distingue el origen de un comando.
    - **Servidor autoritativo**: el cliente solo envía intenciones; toda validación
      (espacio, fondos, stock, plazos) ocurre server-side.
    - **REST para operaciones no urgentes**; los eventos en tiempo real del área de interés
      (movimiento de vehículos, alertas configuradas) se distribuyen por **WebSocket**
      a través del Notification/Event Gateway y quedan fuera de este documento.
    - **El tablón global es pull, no push**: se consulta bajo demanda con filtros
      (producto, ubicación, precio, plazo); nunca hay push mundial de mercado.
    - **Sim-time en el contrato**: todo plazo de dominio (contratos, producción, viajes)
      se define y transmite en sim-time (segundos desde el génesis del mundo, ratio 24×).
      La traducción a tiempo real es responsabilidad exclusiva de la UI. Las dos únicas
      mecánicas en tiempo real son la ventana de sorteo y el cooldown anti-parpadeo.
...
      (invariante del ledger).
...
      distinguibles.
...

    ## Estructura de respuestas

...

...
servers:
...
security:
  - sessionToken: []
...
paths:
...
  /world/concessions:
    get:
      operationId: listConcessions
      tags: [world]
      summary: Concesiones de suelo propias
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ConcessionStatus'
        - name: region_id
          in: query
          schema:
            $ref: '#/components/schemas/RegionId'
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Concesiones de la corporación.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Concession'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/Maintenance' }
    post:
      operationId: createConcession
      tags: [world]
      summary: Obtener una concesión de suelo
      description: |
        Todo terreno es **concesión renovable del sistema** (plazo de referencia: 90 días
        de juego) — no existe propiedad perpetua del suelo (GDD 11.1). El canon periódico
        es un sink estructural; su cuantía deriva del `canon_base` regional ajustado por
        ubicación. El primer canon se cobra al conceder.
      parameters:
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConcessionCreate'
      responses:
        '201':
          description: Concesión otorgada; primer canon asentado en el ledger.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    $ref: '#/components/schemas/Concession'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409':
          description: La parcela solicitada se solapa con una concesión vigente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/Maintenance' }
  /world/concessions/{concessionId}:
    get:
      operationId: getConcession
      tags: [world]
      summary: Detalle de una concesión
      parameters:
        - name: concessionId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/ConcessionId'
      responses:
        '200':
          description: Concesión.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    $ref: '#/components/schemas/Concession'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/Maintenance' }
  /world/concessions/{concessionId}/renew:
    post:
      operationId: renewConcession
      tags: [world]
      summary: Renovar una concesión
      description: Extiende la concesión otro periodo pagando el canon vigente (asiento `canon` del ledger).
      parameters:
        - name: concessionId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/ConcessionId'
        - $ref: '#/components/parameters/idempotencyKey'
      responses:
        '200':
          description: Concesión renovada.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    $ref: '#/components/schemas/Concession'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: La concesión ya está revertida al sistema.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Saldo insuficiente para el canon (`INSUFFICIENT_FUNDS`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/Maintenance' }
  /world/concession-transfers:
    post:
      operationId: createConcessionTransfer
      tags: [world]
      summary: Traspasar una concesión a otra corporación
      description: Mercado secundario de traspasos (GDD 11.1); el sistema cobra una tasa (`system_fee`).
      parameters:
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConcessionTransferCreate'
      responses:
        '201':
          description: Traspaso ejecutado.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    $ref: '#/components/schemas/ConcessionTransfer'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalError' }
        '503': { $ref: '#/components/responses/Maintenance' }

  # ───────────────────────── WORLD: EDIFICIOS Y PRODUCCIÓN ─────────────────────────
...
components:
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
      description: |
        Token de sesión emitido por `POST /auth/sessions`. El servidor solo almacena su
        hash. Humanos y bots usan el mismo esquema (credenciales por cuenta); los bots
        acceden por un camino de red interno multiplexado con la misma superficie de API.
  parameters:
    idempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Clave de idempotencia para **reintentos seguros de comandos que mueven
        valor**: si el cliente reintenta con la misma clave, el servidor reproduce
        la misma respuesta del primer intento — nunca hay doble ejecución.
      schema:
        type: string
        format: uuid
    cursor:
      name: cursor
      in: query
      description: Cursor opaco de paginación devuelto en `meta.next_cursor`.
      schema:
        type: string
    limit:
      name: limit
      in: query
      description: Tamaño máximo de página.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
  responses:
    BadRequest:
      description: Comando malformado o unidades inválidas.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Sesión ausente o expirada.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: Comandar recursos de otra corporación (autorización por propiedad), o vehículo SELLADO durante handoff.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Entidad inexistente (UUID no resuelto).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: Conflicto de estado (aceptación sobre publicación agotada, cooldown, stock ya reservado).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ValidationError:
      description: Error de validación de dominio (422).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    TooManyRequests:
      description: Rate limit excedido — **idéntico para humanos y bots** (decisión de balance, no solo de protección).
      headers:
        Retry-After:
          description: Segundos hasta poder reintentar.
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    InternalError:
      description: Error inesperado. Nunca deja garantías a medio bloquear — toda operación sobre dinero/stock es atómica en el ledger (todo-o-nada).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Maintenance:
      description: |
        Servicio temporalmente no disponible; SIEMPRE reintentable con la MISMA petición.
        Dos causas, distinguibles por el `code`:

        - `MAINTENANCE_WINDOW`: ventana de mantenimiento diaria (sim-time congelado de forma coordinada).
        - `SERIALIZATION_CONFLICT`: la operación chocó con otras concurrentes sobre las mismas
          cuentas del ledger y agotó su presupuesto de reintentos en el servidor. La transacción se
          revirtió ENTERA —no queda ningún efecto parcial, ni garantías a medio bloquear— así que
          reenviar la misma petición tras `Retry-After` es seguro y es lo esperado del cliente.
      headers:
        Retry-After:
          description: Segundos estimados hasta poder reintentar (fin de la ventana, o drenaje de la ráfaga de escrituras).
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            maintenance:
              value:
                error:
                  code: MAINTENANCE_WINDOW
                  message: El mundo está en su ventana de mantenimiento diaria; el sim-time está congelado
                  details: { retry_after_seconds: 900 }
            serialization_conflict:
              value:
                error:
                  code: SERIALIZATION_CONFLICT
                  message: la operación chocó con otras concurrentes y no se aplicó; reintenta la misma petición
                  details: { retry_after_seconds: 1 }
  schemas:
...
    MoneyAmount:
      type: string
      pattern: '^[0-9]+$'
      description: Importe monetario en unidades menores (entero/punto fijo serializado como string — nunca floats; invariante del ledger).
      examples: ["125000"]
...
    SimTime:
      type: integer
      format: int64
      minimum: 0
      description: Segundos de sim-time desde el génesis del mundo (ratio 24× — 1 día de juego = 1 hora real). Todo plazo de dominio se expresa en sim-time.
      examples: [31104000]
...
    AccountId:
      type: string
      format: uuid
      description: UUIDv7 de corporación.
      examples: [01981c5e-7d2a-7f3b-9e41-a2c4d6e8f012]
...
    RegionId:
      type: string
      format: uuid
      description: UUIDv7 de región.
...
    ConcessionId:
      type: string
      format: uuid
      description: UUIDv7 de concesión de suelo.
    ConcessionTransferId:
      type: string
      format: uuid
      description: UUIDv7 de traspaso de concesión.
...
    ConcessionStatus:
      type: string
      enum: [active, delinquent, grace, reverted]
...
    GeoPolygon:
      type: object
      description: Polígono GeoJSON-like (anillo exterior cerrado); vértices en coordenadas planas de mundo en metros `[x_m, y_m]` (SRID 0).
      required: [type, coordinates]
      properties:
        type:
          type: string
          const: Polygon
        coordinates:
          type: array
          items:
            type: array
            minItems: 4
            items:
              type: array
              items:
                type: number
              minItems: 2
              maxItems: 2

    # ── Envoltorios ──
    Meta:
      type: object
      description: Metadatos comunes a toda respuesta exitosa.
      required: [sim_time, server_time]
      properties:
        sim_time:
          type: string
          pattern: '^[0-9]{1,4}-[0-9]{3}-[0-9]{2}:[0-9]{2}$'
          description: 'Sim-time actual del mundo en formato legible `AÑO-DÍA-HH:MM` (p. ej. año 360 de juego, día 45).'
          examples: ["360-045-12:30"]
        sim_time_seconds:
          $ref: '#/components/schemas/SimTime'
          description: Sim-time actual en segundos desde el génesis (forma canónica para clientes).
        server_time:
          type: string
          format: date-time
          description: Wall-clock del servidor (solo informativo — ninguna regla de juego depende de él).
        next_cursor:
          type: string
          description: Cursor de la página siguiente en respuestas paginadas; ausente si no hay más resultados.
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            Código estable de error de dominio. Valores documentados:
            `INSUFFICIENT_COLLATERAL`, `INSUFFICIENT_FUNDS`, `PUBLICATION_EXHAUSTED`,
            `CANCEL_COOLDOWN_ACTIVE`, `BELOW_MIN_LOT`, `STOCK_ALREADY_RESERVED`,
            `PLACEMENT_INVALID`, `NOT_RESOURCE_OWNER`, `VEHICLE_SEALED`,
            `NO_ROUTE_FOUND`, `MAINTENANCE_WINDOW`, `SERIALIZATION_CONFLICT`,
            `RATE_LIMITED`, `VALIDATION_ERROR`, `NOT_FOUND`, `UNAUTHORIZED`,
            `INTERNAL`.
          examples: [INSUFFICIENT_COLLATERAL]
        message:
          type: string
          description: Descripción legible del error.
        details:
          type: object
          additionalProperties: true
          description: Contexto estructurado del error (importes como strings de punto fijo).
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/Error'
      examples:
        - error:
            code: INSUFFICIENT_COLLATERAL
            message: La garantía disponible no cubre la publicación solicitada
            details: { required: "1000", available: "740" }

    # ── Auth ──
...
    Concession:
      type: object
      required: [id, region_id, holder_account_id, parcel, canon_amount, period_sim_days, expires_at_sim, status, granted_at_sim]
      properties:
        id:
          $ref: '#/components/schemas/ConcessionId'
        region_id:
          $ref: '#/components/schemas/RegionId'
        holder_account_id:
          $ref: '#/components/schemas/AccountId'
        parcel:
          $ref: '#/components/schemas/GeoPolygon'
        canon_amount:
          $ref: '#/components/schemas/MoneyAmount'
          description: Canon del periodo (sink estructural; ajustable por el Balancer dentro de rangos).
        period_sim_days:
          type: integer
          minimum: 1
          default: 90
        expires_at_sim:
          $ref: '#/components/schemas/SimTime'
        status:
          $ref: '#/components/schemas/ConcessionStatus'
        granted_at_sim:
          $ref: '#/components/schemas/SimTime'
    ConcessionCreate:
      type: object
      required: [region_id, parcel]
      properties:
        region_id:
          $ref: '#/components/schemas/RegionId'
        parcel:
          $ref: '#/components/schemas/GeoPolygon'
    ConcessionTransferCreate:
      type: object
      required: [concession_id, to_account_id, price]
      properties:
        concession_id:
          $ref: '#/components/schemas/ConcessionId'
        to_account_id:
          $ref: '#/components/schemas/AccountId'
        price:
          $ref: '#/components/schemas/MoneyAmount'
    ConcessionTransfer:
      type: object
      required: [id, concession_id, from_account_id, to_account_id, price, system_fee, occurred_at_sim]
      properties:
        id:
          $ref: '#/components/schemas/ConcessionTransferId'
        concession_id:
          $ref: '#/components/schemas/ConcessionId'
        from_account_id:
          $ref: '#/components/schemas/AccountId'
        to_account_id:
          $ref: '#/components/schemas/AccountId'
        price:
          $ref: '#/components/schemas/MoneyAmount'
        system_fee:
          $ref: '#/components/schemas/MoneyAmount'
        occurred_at_sim:
          $ref: '#/components/schemas/SimTime'

    # ── World: edificios y producción ──
...